Move wrappers.py to axes/plot.py, prepare to add to classes
