+++ docs/1dplots.py
@@ -40,7 +40,7 @@
 # It is often desirable to use different `property cycles
 # <https://matplotlib.org/tutorials/intermediate/color_cycle.html#sphx-glr-tutorials-intermediate-color-cycle-py>`__
 # for different axes or different plot elements. To enable this, the
-# `~proplot.wrappers.cycle_changer` adds the `cycle` and `cycle_kw` to the 1D
+# `~proplot.axes.cycle_changer` adds the `cycle` and `cycle_kw` to the 1D
 # plotting methods. These arguments are passed to the
 # `~proplot.constructor.Cycle` constructor function, and the resulting property
 # cycle is used to style the input data. ProPlot iterates through property
@@ -76,9 +76,9 @@ with plot.rc.context({'lines.linewidth': 3}):
 # Standardized arguments
 # ----------------------
 #
-# The `~proplot.wrappers.standardize_1d` wrapper is used to standardize
+# The `~proplot.axes.standardize_1d` wrapper is used to standardize
 # positional arguments across all 1D plotting methods.
-# `~proplot.wrappers.standardize_1d` allows you to optionally omit *x*
+# `~proplot.axes.standardize_1d` allows you to optionally omit *x*
 # coordinates, in which case they are inferred from the data. It also permits
 # passing 2D *y* coordinate arrays to any plotting method, in which case the
 # plotting method is called for each column of the array.
@@ -122,7 +122,7 @@ with plot.rc.context({'axes.prop_cycle': plot.Cycle('Grays', N=N, left=0.3)}):
 # Pandas and xarray integration
 # -----------------------------
 #
-# The `~proplot.wrappers.standardize_1d` wrapper integrates 1D plotting
+# The `~proplot.axes.standardize_1d` wrapper integrates 1D plotting
 # methods with pandas `~pandas.DataFrame`\ s and xarray `~xarray.DataArray`\ s.
 # When you pass a DataFrame or DataArray to any plotting command, the x-axis
 # label, y-axis label, legend label, colorbar label, and/or title are
@@ -185,7 +185,7 @@ axs[1].plot(df, cycle=cycle, lw=3, legend='uc')
 # Adding error bars
 # -----------------
 #
-# The `~proplot.wrappers.add_errorbars` wrapper lets you draw error bars
+# The `~proplot.axes.add_errorbars` wrapper lets you draw error bars
 # on-the-fly by passing certain keyword arguments to
 # `~matplotlib.axes.Axes.plot`, `~matplotlib.axes.Axes.scatter`,
 # `~matplotlib.axes.Axes.bar`, or `~matplotlib.axes.Axes.barh`.
@@ -193,12 +193,12 @@ axs[1].plot(df, cycle=cycle, lw=3, legend='uc')
 # If you pass 2D arrays to these methods with ``means=True`` or
 # ``medians=True``, the means or medians of each column are drawn as points,
 # lines, or bars, and error bars are drawn to represent the spread in each
-# column. `~proplot.wrappers.add_errorbars` lets you draw both thin error
+# column. `~proplot.axes.add_errorbars` lets you draw both thin error
 # "bars" with optional whiskers, and thick error "boxes" overlayed on top of
 # these bars (this can be used to represent different percentil ranges).
 # Instead of using 2D arrays, you can also pass error bar coordinates
 # *manually* with the `bardata` and `boxdata` keyword arguments. See
-# `~proplot.wrappers.add_errorbars` for details.
+# `~proplot.axes.add_errorbars` for details.
 
 # %%
 import proplot as plot
@@ -254,8 +254,8 @@ plot.rc.reset()
 # ------------------------
 #
 # The `~matplotlib.axes.Axes.bar` and `~matplotlib.axes.Axes.barh` methods
-# are wrapped by `~proplot.wrappers.bar_wrapper`,
-# `~proplot.wrappers.cycle_changer`, and `~proplot.wrappers.standardize_1d`.
+# are wrapped by `~proplot.axes.bar_wrapper`,
+# `~proplot.axes.cycle_changer`, and `~proplot.axes.standardize_1d`.
 # You can now *group* or *stack* columns of data by passing 2D arrays to
 # `~matplotlib.axes.Axes.bar` or `~matplotlib.axes.Axes.barh`, just like in
 # `pandas`. Also, `~matplotlib.axes.Axes.bar` and `~matplotlib.axes.Axes.barh`
@@ -266,8 +266,8 @@ plot.rc.reset()
 # `~proplot.axes.Axes.areax` methods. These are alises for
 # `~matplotlib.axes.Axes.fill_between` and
 # `~matplotlib.axes.Axes.fill_betweenx`, which are now wrapped by
-# `~proplot.wrappers.fill_between_wrapper` and
-# `~proplot.wrappers.fill_betweenx_wrapper`. You can now *stack* or *overlay*
+# `~proplot.axes.fill_between_wrapper` and
+# `~proplot.axes.fill_betweenx_wrapper`. You can now *stack* or *overlay*
 # columns of data by passing 2D arrays to `~proplot.axes.Axes.area` and
 # `~proplot.axes.Axes.areax`, just like in `pandas`. You can also now draw
 # area plots that *change color* when the fill boundaries cross each other by
@@ -350,9 +350,9 @@ plot.rc.reset()
 # --------------------------
 #
 # The `~matplotlib.axes.Axes.boxplot` and `~matplotlib.axes.Axes.violinplot`
-# methods are now wrapped with `~proplot.wrappers.boxplot_wrapper`,
-# `~proplot.wrappers.violinplot_wrapper`, `~proplot.wrappers.cycle_changer`,
-# and `~proplot.wrappers.standardize_1d`. These wrappers add some useful
+# methods are now wrapped with `~proplot.axes.boxplot_wrapper`,
+# `~proplot.axes.violinplot_wrapper`, `~proplot.axes.cycle_changer`,
+# and `~proplot.axes.standardize_1d`. These wrappers add some useful
 # options and apply aesthetically pleasing default settings. They also
 # automatically apply axis labels based on the `~pandas.DataFrame` column
 # labels or the input *x* coordinate labels.
@@ -456,8 +456,8 @@ ax.colorbar(m, loc='b', maxn=10, label=f'parametric coordinate')
 # -------------
 #
 # The `~matplotlib.axes.Axes.scatter` method is now wrapped by
-# `~proplot.wrappers.scatter_wrapper`, `~proplot.wrappers.cycle_changer`, and
-# `~proplot.wrappers.standardize_1d`. This means that
+# `~proplot.axes.scatter_wrapper`, `~proplot.axes.cycle_changer`, and
+# `~proplot.axes.standardize_1d`. This means that
 # `~matplotlib.axes.Axes.scatter` now accepts 2D arrays, just like
 # `~matplotlib.axes.Axes.plot`. Also, successive calls to
 # `~matplotlib.axes.Axes.scatter` now use the property cycler properties
