+++ docs/colormaps.py
@@ -139,7 +139,7 @@ for cmaps in (('magma', 'rocket'), ('fire', 'dusk')):
 # your convenience, most of these features can be accessed via the
 # `~proplot.constructor.Colormap` constructor function. Note that every
 # plotting command that accepts a `cmap` keyword passes it through this
-# function (see `~proplot.wrappers.cmap_changer`).
+# function (see `~proplot.axes.cmap_changer`).
 #
 # To make `~proplot.colors.PerceptuallyUniformColormap`\ s from scratch, you
 # have the following three options:
