+++ proplot/utils.py
@@ -93,7 +93,7 @@ def edges(Z, axis=-1):
     you supply centers to `~matplotlib.axes.Axes.pcolor` or
     `~matplotlib.axes.Axes.pcolormesh`. It is also used to calculate colormap
     level boundaries when you supply centers to plotting methods wrapped by
-    `~proplot.wrappers.cmap_changer`.
+    `~proplot.axes.cmap_changer`.
 
     Parameters
     ----------
